wrapped the code in a lambda function. Expects a "data" String.
